+++ datacube/drivers/postgis/_api.py
@@ -17,11 +17,19 @@ import datetime
 import json
 import logging
 import uuid  # noqa: F401
-from sqlalchemy import cast
-from sqlalchemy import delete, update
+from sqlalchemy import (
+    cast,
+    delete,
+    update,
+    select,
+    text,
+    and_,
+    or_,
+    func,
+    column,
+)
 from sqlalchemy.dialects.postgresql import insert
 from sqlalchemy.sql.expression import Select
-from sqlalchemy import select, text, and_, or_, func
 from sqlalchemy.dialects.postgresql import INTERVAL
 from sqlalchemy.exc import IntegrityError
 
@@ -50,6 +58,34 @@ _LOG = logging.getLogger(__name__)
 # Make a function because it's broken
 def _dataset_select_fields() -> tuple:
     return tuple(f.alchemy_expression for f in _dataset_fields())
+
+
+def _base_known_fields():
+    fields = get_native_fields().copy()
+    fields['archived'] = NativeField(
+        'archived',
+        'Archived date',
+        Dataset.archived
+    )
+    fields["uris"] = NativeField(
+        "uris",
+        "all uris",
+        func.array(
+            select(
+                SelectedDatasetLocation.uri
+            ).where(
+                and_(
+                    SelectedDatasetLocation.dataset_ref == Dataset.id,
+                    SelectedDatasetLocation.archived == None
+                )
+            ).order_by(
+                SelectedDatasetLocation.added.desc(),
+                SelectedDatasetLocation.id.desc()
+            ).label('uris')
+        ),
+        alchemy_table=Dataset.__table__  # type: ignore[attr-defined]
+    )
+    return fields
 
 
 def _dataset_fields() -> tuple:
@@ -634,13 +670,14 @@ class PostgisDbAPI:
     def search_datasets_query(self,
                               expressions, source_exprs=None,
                               select_fields=None, with_source_ids=False, limit=None, geom=None,
-                              archived: bool | None = False):
+                              archived: bool | None = False, order_by=None):
         """
         :type expressions: Tuple[Expression]
         :type source_exprs: Tuple[Expression]
         :type select_fields: Iterable[PgField]
         :type with_source_ids: bool
         :type limit: int
+        :type order_by: Iterable[Any]
         :type geom: Geometry
         :rtype: sqlalchemy.Expression
         """
@@ -650,6 +687,24 @@ class PostgisDbAPI:
 
         if not select_fields:
             select_fields = _dataset_fields()
+
+        known_fields = _base_known_fields() | {f.name: f for f in select_fields}
+
+        def _ob_exprs(o):
+            if isinstance(o, str):
+                if known_fields.get(o.lower()) is not None:
+                    return known_fields[o.lower()].alchemy_expression
+                raise ValueError(f"Cannot order by unknown field {o}")
+            elif isinstance(o, PgField):
+                return o.alchemy_expression
+            else:
+                # assume func, clause, or other expression, and leave as-is
+                return o
+
+        if order_by is not None:
+            order_by = [_ob_exprs(o) for o in order_by]
+        else:
+            order_by = []
 
         select_columns = tuple(
             f.alchemy_expression.label(f.name)
@@ -678,13 +733,13 @@ class PostgisDbAPI:
         if spatialquery is not None:
             where_expr = and_(where_expr, spatialquery)
             query = query.join(SpatialIndex)
-        query = query.where(where_expr).limit(limit)
+        query = query.where(where_expr).order_by(*order_by).limit(limit)
         return query
 
     def search_datasets(self, expressions,
                         source_exprs=None, select_fields=None,
                         with_source_ids=False, limit=None, geom=None,
-                        archived: bool | None = False):
+                        archived: bool | None = False, order_by=None):
         """
         :type with_source_ids: bool
         :type select_fields: tuple[datacube.drivers.postgis._fields.PgField]
@@ -692,11 +747,12 @@ class PostgisDbAPI:
 
         :return: An iterable of tuples of decoded values
         """
+        assert source_exprs is None
+        assert not with_source_ids
         if select_fields is None:
             select_fields = _dataset_fields()
-        select_query = self.search_datasets_query(expressions, source_exprs,
-                                                  select_fields, with_source_ids,
-                                                  limit, geom=geom, archived=archived)
+        select_query = self.search_datasets_query(expressions, select_fields=select_fields, limit=limit,
+                                                  geom=geom, archived=archived, order_by=order_by)
         _LOG.debug("search_datasets SQL: %s", str(select_query))
 
         def decode_row(raw: Iterable[Any]) -> dict[str, Any]:
@@ -1541,3 +1597,18 @@ class PostgisDbAPI:
         return select(
             func.min(time_min.alchemy_expression), func.max(time_max.alchemy_expression)
         )
+
+    def find_most_recent_change(self, product_id: int):
+        """
+        Find the database-local time of the last dataset that changed for this product.
+        """
+        return self._connection.execute(
+            select(
+                func.max(
+                    func.greatest(
+                        Dataset.added,
+                        column("updated"),
+                    )
+                )
+            ).where(Dataset.product_ref == product_id)
+        ).scalar()