+++ datacube/index/postgis/_products.py
@@ -385,3 +385,11 @@ class ProductResource(AbstractProductResource, IndexResourceAddIn):
         ids = [ds.id for ds in self._index.datasets.search(product=product.name)]
         with self._db_connection() as connection:
             return connection.spatial_extent(ids, crs)
+
+    def most_recent_change(self, product: str | Product) -> datetime.datetime | None:
+        if isinstance(product, str):
+            product = self._index.products.get_by_name_unsafe(product)
+        assert isinstance(product, Product)
+        assert product.id is not None
+        with self._db_connection() as connection:
+            return connection.find_most_recent_change(product.id)