+++ docs/about/whats_new.rst
@@ -20,6 +20,7 @@ v1.9.next
 - Remove problematic "common_timestamp" postgresql function from postgis driver. Some internal API changes
   required to accommodate and preserve all previous index-driver level behaviour. (:pull:`1623`)
 - Cherry picks from 1.8 (#1624-#1626, #1629-#1631) (:pull:`1637`)
+- Updates to index APIs - add `order_by` to dataset search, index name attribute, product most recent change method (:pull:`1643`)
 - Fix alembic migrations (:pull:`1645`)
 
 v1.9.0-rc9 (3rd July 2024)